refactor(pad): extract shared pad_validate + pad_resolve core
Move the pad JSON validator out of mcp_tools_pads.cpp into a standalone
pad_validate.{h,cpp} (gated HAS_DISPLAY, independent of HAS_MCP) so the web
portal save path and the MCP tools can validate through identical logic
instead of the current thinner portal check. No behavior change for MCP.

Add pad_resolve() to pad_binding: batch-resolve N binding templates against
an optional pad context with save/restore, powering the upcoming MCP
resolve_bindings tool and the portal /api/pad/resolve preview. Host test
added for the batch helper.
